+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Support for parameter declarations.
 - Support for typedefs.
+- Support for port assignments in instantiations.
 
 ### Fixed
 - Fix assigning values to output ports.